backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
requests==2.26.0
supabase==1.0.3
python-dateutil==2.8.2
pytz==2021.1 